services/agents/requirements.txt
# System packages
## ffmpeg
## espeak

transformers==4.52.4
datasets[audio]==3.6.0
accelerate==1.7.0
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121
--extra-index-url https://download.pytorch.org/whl/cu121
fastapi==0.115.9
uvicorn==0.34.2
pydub==0.25.1
soundfile==0.13.1
openai-whisper==20240930
numpy==2.2.6
PyPDF2==3.0.1
chromadb==1.0.12
flask

SpeechRecognition==3.14.3
langchain[openai]==0.3.25
langchain_community==0.3.24
langchain_huggingface==0.2.0
langchain_chroma==0.2.4
pyttsx3==2.98
